refactor: split test_policy.py into focused per-module test files
Split the 515-line grab-bag test_policy.py into 3 files that map 1:1
to the source modules they test:

- test_policy_parser.py: flatten, DefaultContext, placeholders, rule_to_dict
- test_policy_matcher.py: YAML fixture match verdicts, unknown attributes
- test_policy_enforcer.py: YAML fixture enforcer scenarios

Also moved TestDumpRules from test_policy_cli.py to test_policy_parser.py
since it tests parse_policy + rule_to_dict (parser functions, not CLI).

779 tests still pass.

Closes #127
